requirements.txt
# LangChain & LangGraph
langchain>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
langchain-chroma>=0.2.0

# Vector Store
chromadb>=0.5.0

# UI
gradio>=5.0.0

# Ollama
ollama>=0.3.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
aiohttp>=3.9.0
nest-asyncio>=1.6.0

# Web scraping
beautifulsoup4>=4.12.0
requests>=2.31.0

# Document processing
markdown>=3.5.0
pypdf>=4.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
